+++ src/Tests/Dfe.Complete.CypressTests/cypress/api/taskApiConversions.ts
@@ -55,22 +55,18 @@ interface UpdateDirectionToTransferTaskRequest {
 
 interface UpdateLandQuestionnaireTaskRequest {
     taskDataId: TaskDataId;
-    received?: boolean;
-    cleared?: boolean;
-    signed?: boolean;
-    saved?: boolean;
-}
-
-interface UpdateLandRegistryTitlePlansTaskRequest {
-    taskDataId: TaskDataId;
-    received?: boolean;
-    cleared?: boolean;
-    saved?: boolean;
+    landQuestionnaireReceived?: boolean;
+    landQuestionnaireCleared?: boolean;
+    landQuestionnaireSigned?: boolean;
+    landQuestionnaireSaved?: boolean;
+    landRegistryTitlePlansReceived?: boolean;
+    landRegistryTitlePlansCleared?: boolean;
+    landRegistryTitlePlansSaved?: boolean;
 }
 
 interface UpdateOneHundredAndTwentyFiveYearLeaseTaskRequest {
     taskDataId: TaskDataId;
-    notApplicable?: boolean;
+    confirm?: boolean;
     email?: boolean;
     receive?: boolean;
     save?: boolean;
@@ -187,10 +183,6 @@ class TaskApiConversions extends TaskApi {
 
     public updateLandQuestionnaireTask(requestBody: UpdateLandQuestionnaireTaskRequest) {
         return this.taskDataBaseRequest<void>("LandQuestionnaire", requestBody);
-    }
-
-    public updateLandRegistryTitlePlansTask(requestBody: UpdateLandRegistryTitlePlansTaskRequest) {
-        return this.taskDataBaseRequest<void>("LandRegistryTitlePlans", requestBody);
     }
 
     public updateOneHundredAndTwentyFiveYearLeaseTask(requestBody: UpdateOneHundredAndTwentyFiveYearLeaseTaskRequest) {