+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Security in case of vulnerabilities.
 
 ### Changed
  - [279036] (https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/279036) - Direction to transfer content changes
+ - [277534](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/277534) - Merge conversion tasks "Land questionnaire" and "Land registry title plans" into "Land questionnaire(s) and land registry plans" with merged page/handler updates
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2026-05-21.1562...main) for everything awaiting release
 
@@ -26,6 +27,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - [277849](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/277849) - Update wording on articles of association conversion tasks
 - [277820](https://dev.azure.com/dfe-gov-uk/Academies-and-Free-Schools-SIP/_workitems/edit/277820) - Church Supplemental Agreement
 - [282906](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/282906) - Update wording on confirm academy open date
+- [279037](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/279037) - updated checkboxes on 125 year lease
 
 ### Added
 - [279042](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_workitems/edit/279042) - add conversion task "Private finance initiative" with GET and POST support, task list status, migration, and test coverage